pyproject.toml
[tool.poetry]
package-mode = false
name = "pygame-template-renewed"
description = "Pygame template."
authors = ["DaNubCoding <danubcoding@gmail.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame-ce = "^2.5.2"
multimethod = "^1.12"

[tool.poe.tasks]
run = "python main.py"
new-scene = "python scripts/create_scene.py"
new-sprite = "python scripts/create_sprite.py"
init-debug = "python scripts/init_debug.py"
